requirements.txt
python-dotenv
yt-dlp
openai-whisper
tiktoken
openai 
